+++ kitty/config_data.py
@@ -964,9 +964,15 @@ o('macos_traditional_fullscreen', False, long_text=_('''
 Use the traditional full-screen transition, that is faster, but less pretty.
 '''))
 
-o('macos_show_window_title_in_menubar', True, long_text=_('''
-Show the title of the currently active window in the macOS
-menu-bar, making use of otherwise wasted space.'''))
+o('macos_show_window_title_in', 'all', option_type=choices('all', 'window', 'menubar', 'none'), long_text=_('''
+Show or hide the window title in the macOS window or menu-bar.
+A value of :code:`window` will show the title of the currently
+active window at the top of the macOS window. A value of
+:code:`menubar` will show the title of the currently active window
+in the macOS menu-bar, making use of otherwise wasted space.
+:code:`all` will show the title everywhere and :code:`none`
+hides the title in the window and the menu-bar.
+'''))
 
 # Disabled by default because of https://github.com/kovidgoyal/kitty/issues/794
 o('macos_custom_beam_cursor', False, long_text=_('''